fix(wallet): address 9.2 code-review findings (settlement & coin UX)
- record true wallet deltas on settlement credit failure (all humans -buyIn,
  winners uncredited) instead of optimistic computed deltas
- tighten event:coin_settlement coin fields to integers (Zod .int() + pot guard)
- fix join-by-code INSUFFICIENT_COINS toast to use the param-less generic key

3-layer adversarial code review: 3 patches applied, 6 items deferred to
Story 9.3 / crash-recovery, 8 dismissed. Story 9.2 -> done.

+++ client/src/shared/types/wsEvents.schemas.ts
@@ -265,9 +265,9 @@ export const AutoActionPayloadSchema = z.strictObject({
 
 // Story 9.2: per-human match-end coin settlement.
 export const CoinSettlementPayloadSchema = z.strictObject({
-  coinDelta: z.number(),
-  newBalance: z.number(),
-  pot: z.number(),
+  coinDelta: z.number().int(),
+  newBalance: z.number().int(),
+  pot: z.number().int(),
 });
 
 export const PlayerDisconnectedPayloadSchema = z.strictObject({
